[#5] refactor: model페키지 삭제
