+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/feed/entity/FeedLike.java
@@ -6,7 +6,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.goormthon.seasonthon.nocheongmaru.domain.member.model.entity.Member;
+import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Member;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
