+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/feed/repository/feed/FeedRepository.java
@@ -2,7 +2,6 @@ package org.goormthon.seasonthon.nocheongmaru.domain.feed.repository.feed;
 
 import static org.goormthon.seasonthon.nocheongmaru.domain.feed.entity.QFeed.*;
 import static org.goormthon.seasonthon.nocheongmaru.domain.feed.entity.QFeedLike.*;
-import static org.goormthon.seasonthon.nocheongmaru.domain.member.model.entity.QMember.*;
 
 import java.util.List;
 import java.util.Optional;
