+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/information/entity/Information.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.goormthon.seasonthon.nocheongmaru.domain.member.model.entity.Member;
+import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Member;
 import org.goormthon.seasonthon.nocheongmaru.global.common.BaseTimeEntity;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
