+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/member/repository/MemberJpaRepository.java
@@ -1,7 +1,7 @@
 package org.goormthon.seasonthon.nocheongmaru.domain.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.goormthon.seasonthon.nocheongmaru.domain.member.model.entity.Member;
+import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Member;
 
 import java.util.Optional;
