+++ src/test/java/org/goormthon/seasonthon/nocheongmaru/global/security/jwt/provider/TokenProviderTest.java
@@ -1,7 +1,7 @@
 package org.goormthon.seasonthon.nocheongmaru.global.security.jwt.provider;
 
 import org.goormthon.seasonthon.nocheongmaru.IntegrationTestSupport;
-import org.goormthon.seasonthon.nocheongmaru.domain.member.model.entity.Role;
+import org.goormthon.seasonthon.nocheongmaru.domain.member.entity.Role;
 import org.goormthon.seasonthon.nocheongmaru.global.security.jwt.dto.TokenResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
